python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-shell"
version = "0.1.0"
description = "An interactive shell for AWS with rich tables, tab completion, and a Python REPL"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.34.0",
    "prompt-toolkit>=3.0.43",
    "Pygments>=2.17.0",
    "rich>=13.7.0",
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
]

[project.scripts]
aws-shell = "aws_shell.cli:main"

[tool.setuptools.packages.find]
include = ["aws_shell*"]
